zdaemon/requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cowsay==6.1
distro==1.9.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
slack-bolt==1.18.1
slack_sdk==3.27.0
typing_extensions==4.9.0
urllib3==2.1.0
zulip==0.9.0
